extract only the page template modal
