+++ packages/page-template-modal/package.json
@@ -35,7 +35,6 @@
 		"@wordpress/data": "^4.26.1",
 		"@wordpress/element": "^2.19.01",
 		"@wordpress/blocks": "^6.25.1",
-		"@wordpress/hooks": "^2.11.0",
 		"@wordpress/api-fetch": "^3.3.0",
 		"@wordpress/url": "^2.21.0",
 		"lodash": "^4.17.19",
